+++ api/routes/posts.js
@@ -48,71 +48,4 @@ router.get('/:post_id', (req, res, next) => {
   }).catch(err => next(err));
 });
 
-
-/** 
- * COMMENTS COMMENTS COMMENTS COMMENTS
- */
-
-router.get('/:post_id/comments', (req, res, next) => {
-  const postId = parseInt(req.params['post_id']);
-  Comments.getPostComments(postId).then(result => {
-    res.status(200).send(result.rows)
-  }).catch(err => next(err));
-});
-
-
-router.post('/:post_id/comments', (req, res, next) => {
-  const postId = parseInt(req.params['post_id']);
-  const { user_id, text } = req.body;
-  Comments.addPostComment(postId, user_id, text).then(result => {
-    res.status(201).send({message: `created comment with id ${result}`})
-  }).catch(err => next(err));
-});
-
-
-router.delete('/:post_id/comments', (req, res, next) => {
-  const postId = parseInt(req.params['post_id']);
-  const commentId = parseInt(req.query['comment_id']);
-  Comments.deletePostComment(postId, commentId).then(result => {
-    res.status(200).send({message: 'comment deleted'});
-  }).catch(err => next(err));
-});
-
-
-/**
- * ----- STARS STARS STARS ------
- */
- router.get('/:post_id/stars', (req, res, next) => {
-   const postId = parseInt(req.params['post_id']);
-   Stars.getPostStars(postId).then(result => {
-     res.status(200).send(result.rows)
-   }).catch(err => next(err));
- });
-
-
- router.get('/:post_id/star_count', (req, res, next) => {
-  const postId = parseInt(req.params['post_id']);
-  Stars.getPostStarCount(postId).then(result => {
-    res.status(200).send(result.rows[0])
-  }).catch(err => next(err));
- });
-
-
- router.post('/:post_id/stars', (req, res) => {
-   const postId = parseInt(req.params['post_id']);
-   const userId = parseInt(req.query['user_id']);
-   Stars.addPostStar(userId, postId).then(result => {
-     res.status(201).send({message: 'created star'})
-   }).catch(err => next(err));
- });
-
-
- router.delete('/:post_id/stars', (req, res) => {
-   const postId = parseInt(req.params['post_id']);
-   const userId = parseInt(req.query['user_id']);
-   Stars.deletePostStar(postId, userId).then(result => {
-     res.status(200).send({message: 'star deleted'});
-   }).catch(err => next(err));
- });
-
 module.exports = router;
